+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Change Log
 ==========
 
 ## [Unreleased]
+* Ability to now publish a message without wrapping it in a Message object
 
 
 ## [0.1.0]